fix (ai/core): move process access into functions and use globalThis
