+++ .changeset/metal-impalas-wink.md
@@ -1,0 +1,5 @@
+---
+'ai': patch
+---
+
+fix (ai/core): move process access into functions and use globalThis
